Initial draft of enginized/gemified app generator
Restructures the code into an engine (/core) and suite of gems (answers
and answers-core).  This also marks the first draft of the new app
generator.  There is still a good amount of work left to restore much
of the original functionality of the rails app (e.g. admin panel,
assets pipeline, gem dependencies, tests, and chef/vagrant bits).

To test, run `./bin/answers MyAnswers` from the root of this project.